feat: rpc redis
