+++ packages/brokers/src/index.ts
@@ -1,2 +1,5 @@
+export * from './brokers/redis/BaseRedis';
 export * from './brokers/redis/PubSubRedis';
+export * from './brokers/redis/RPCRedis';
+
 export * from './brokers/Broker.interface';
